Wyrzucony plik z algorytmem, przebudowane testy pod to, żeby nie korzystały z wcześniej błędnie dopisanego przeze mnie algorytmu
